Basic Model Saving implemented
We now have the basic framework necessary to save models upon successfully training a model (which currently is defined by having an epoch that gets a mean score of above 20000). Given that this has never occured thus far it might be a bit superflous, but better early than late. Exceptionb handlibng to allow for patrial model training should be coming soon.