changed to one hot encoding of the board state for passing to the model
In prior versions of this project the model was trained using a 1x16 tensor of the actual values in the 2048 board, however this is believed to have been making it impossible for smaller models to learn good play, and much harder even for large models due to the relative differences between different possible values (ie, 2 and 4 vs 64 and 256). Taking the log of those values would have solved this problem, however we believe one hot endcoding to be a better solution (currently) because it removes the scalar nature of the numbers since we belive it distinguishes the different possible numbers better.